+++ src/__init__.py
@@ -1,3 +1,3 @@
 """Waffler core modules"""
 
-__version__ = "3.12.6"
+__version__ = "3.12.7"